Move lean configuration to single space
