+++ app/config/include/lean.neon
@@ -1,0 +1,23 @@
+leanMapper:
+    db: %database%
+
+leanFixtures:
+    initFiles:
+        - %rootDir%/data/testing/commonModuleFixtures.neon
+    daos:
+        - PAF\Modules\CommonModule\Fixtures\UserFixtureDao()
+        - PAF\Common\Fixtures\RepositoryFixtureDao(@common.personRepository)
+        - PAF\Common\Fixtures\RepositoryFixtureDao(@common.contactRepository)
+
+services:
+    leanMapper.mapper:
+        class: SeStep\ModularLeanMapper\ModularMapper
+        arguments:
+            mapper: PAF\Common\Model\UnderscoreMapper()
+            modules:
+                ss_settings__: SeStep\LeanSettings\LeanOptionsMapperModule()
+                ss_files__: SeStep\FileAttachable\FilesMapperModule()
+                ss_comments__: SeStep\Commentable\Lean\CommentsMapperModule()
+                app_log__: SeStep\ModularLeanMapper\MapperModule::create('PAF\Modules\ApplicationLogModule')
+                common__: SeStep\ModularLeanMapper\MapperModule::create('PAF\Modules\CommonModule', 'Model')
+                commission__: SeStep\ModularLeanMapper\MapperModule::create('PAF\Modules\CommissionModule', 'Model')
